Move ClickHouseStatementWrapper to the connection package and format README.md
