+++ src/main/java/org/apache/flink/connector/clickhouse/internal/converter/ClickHouseRowConverter.java
@@ -1,5 +1,6 @@
 package org.apache.flink.connector.clickhouse.internal.converter;
 
+import org.apache.flink.connector.clickhouse.internal.connection.ClickHouseStatementWrapper;
 import org.apache.flink.table.data.DecimalData;
 import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
